backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
dj-database-url==2.3.0
psycopg2-binary==2.9.9
python-jose==3.3.0
python-dotenv==1.0.1
